+++ CHANGELOG.md
@@ -14,7 +14,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `Cargo.lock` is exactly what is expected (without implicit changes at build
   time) but this may end up rejecting builds which were previously passing. To
   get the old behavior back, set `cargoExtraArgs = "";`
-* Fixed a bug when testing proc macro crates with `cargoNextest` on macO. ([#376](https://github.com/ipetkov/crane/pull/376))
+* **Breaking**: `cargoDoc` will no longer install cargo artifacts by default.
+  Set `doInstallCargoArtifacts = true;` to get the old behavior back.
+* `cargoDoc` will now install generated documentation in `$out/share/doc`
+* Fixed a bug when testing proc macro crates with `cargoNextest` on macOS.
+  ([#376](https://github.com/ipetkov/crane/pull/376))
 
 ## [0.13.1] - 2023-08-22